Moving commands component to component directory
